Use binary IO channel for input_string
The channel encoding defines to which encoding the input string should
be transformed before writing it to the channel. Since the in-memory
encoding is always UTF-8 and the default encoding is also UTF-8, this is
a NO-OP by default — except that it checks that all characters are valid
UTF-8 characters. To avoid this check the encoding must be set to NULL
(=binary). Note that the actual transformation (=NO-OP) remains the
same.