Catch bogus UUID changes of MDRAIDs.
Right before the "remove" event, the UUID of mdraid devices seems to
change sometimes.

Without catching this, we would create a new object for the new UUID
and erroneously dispatch the "remove" event to it.  Consequently, the
original object would not receive this event and would continue
watching a non-existent device.