ISCSI: introduce parameters passing for nodes when Login/Logout is issued (fixes #31)
When the o.s.S.M.ISCSI.Initiator.Login is issued and we would like to
set the node to automatically log in to a portal, it's desired to set
'node.startup' property to 'automatic'.

This can be achieved by passing an optional parameter to Login command:
{'node.startup': <'automatic'>}.

The same approach can be applied for Logout method:
{'node.startup': <'manual'>}.

Fixes rhbz#1265673.