Introduce configuration file for storaged
This configuration file contains options for specifying what modules
should be loaded and when/how they should be loaded. Key 'modules' may
contain comma-separated list of modules (libstoraged_{module_name}.so).
On the other hand, the key 'modules_load_preference' controlls when the
modules will be loaded; either at startup ('onstartup') or when users
requests it ('ondemand' which is the default value).