Core: Survive random remove/add uevent pairs on blocks
These event pairs usually happen asynchronously sometime after
creating a partition, when udev issues a BLKRRPART as part of
synthesizing a change event for the parent device.

This tries to fix the same issue as eb4b994, which doesn't seem to be
effective.  Unfortunately, the fix only helps with the storaged
internals; it is still possible that external programs such as wipefs
will find the block device missing when they happen to look for it
after the remove and before the add.