Revert "Reread partition table before wiping when creating new partitions"
This reverts commit eb4b994c8a12c77744431f1a8459126889c647f9.

The BLKRRPART ioctl is dangerous in that it causes pairs of remove/add
uevents for all partitions which in turn confuse our internal data
structures and cause the block devices to briefly disappear from the
filesystem.

Rereading the partition table is also unnecessary since parted,
libblockdev, and udevd do that already.

Thus, eb4b994c might actually make things worse instead of fixing the
bug it is intended to fix.