Don't coldplug uninitilized udev devices
When the udev device is not initialized it may not contain all of its
properties yet, which makes udisks provide an incorrect representation
of the to its clients. This is particularly problematic during the
system initialization phase, where udev is also in it coldplug phase.

udev will send an event when each device becomes initialized, so we can
safely skip it during coldplug and rely on hotplug even handling
instead.

https://bugs.freedesktop.org/show_bug.cgi?id=96832
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>