Tests: Add basic D-Bus tests for LVM and iSCSI
These are just the basic LVM and iSCSI tests borrowing lot of code and ideas
from the integration_test already present in the tree. The tests are by no means
complete and would need to be extended. They intentionally don't rely on the gi
python module.
The tests are meant to be run on a dedicated machine: they connect to the system
bus and may disrupt the running system.