Core: Force re-applying configuration on system resume from sleep
When the system suspends or hibernates the ATA drives reset their
configuration to default. This for example resets the APM level. This
is an attempt to fix the problem. Storaged subscribes to logind D-Bus
signal sent on system suspending or resuming and on resume event scans
scans the drive configuration files and for each such drive synthesizes
a fake "reconfigure" event that makes the uevent handling routines to
apply the configuration settings again.