Revert "Build: Add srpm make target; autogenerate rpm versions from git hash."
This reverts commit 634bbf6bb00c527b03bc077e0096907e86832840. Some other approach
to automated RPM versioning must be taken. The SRPM must contain "hard-coded"
version so in the case it's been produced from the git tree it would always carry
the git commit information.

Since rpm uses the SPEC file in the SRPM it would produce a differently versioned
binary RPMs when built outside the git tree.