Build: RPM target -- autogenerate the version tag
This is another attempt to improve the RPM build target. The version
and release tags must be autogenerated. In order not to break the
'generate' script in packaging directory, I've added a new 'version.sh'
script which takes the version string from conigure.ac and if run from
within the git tree creates a git hash based release string for rpm.