lvm2daemonutil.c: Close FD for spawned process
When the lvm helper program was being called we leaked the FD that
was used to comunicate with the child process.  I would think that
when you clean-up a channel the utilized FD would get closed, but
it appears not.  Add an explicit close to the FD when we are
cleaning up the spawned process.

Signed-off-by: Tony Asleson <tasleson@redhat.com>