Don't add new drive object on "change" uevent
When hot-unplugging a SATA disk, udev sends a sequence of events: first
a "change" and then "remove". Storaged reacts to the "change" event by
creating a new drive object on the bus. Since there is no vendor, model
and serial device info, the path of the new object is some fallback one
("/org/freedesktop/UDisks2/drives/drive") and this is the object that also
gets deleted by the subsequent "remove" action. The original drive object
is never removed and storage attempts to perform housekeeping on it,
resulting in various errors.

The patch prevents the creation of the new drive object on the "change"
event. It's actually odd to create a new device on such occassion.