Replace g_error_free() with g_clear_error()
The difference between 'g_error_free (error)' and 'g_clear_error (&error)' is
that g_clear_error also sets 'error' to NULL. That may seem like a negligible
difference, but it is a huge one because many GLib (and other) functions have
the assert that 'error' is NULL as the first thing they do returning (NULL) if
the assert fails. So an error we may handle gracefully may cause follow-up
failures just because the variable is not set to NULL (indicating there has been
an unhandled error before).

There doesn't seem to be a single place where the above behaviour would be
desired so this patch replaces all g_error_free() calls by g_clear_error()
ones. It may not be necessary in some places, but I think a few extra
instructions for 'error = NULL' are worth the removal of potentially problematic
function calls from the codebase.