Remove the false argument of StoragedTestCase.get_object()
iface_suffix has nothing to do with any DBus interface. The value is appended to
the "org.freedesktop.UDisks2" and used as the first argument for the
bus.get_object() method which is 'bus_name' (the name acquired by the
service). That's always "org.freedesktop.UDisks2" in all of our tests so the
argument to our simplified get_object() method just doesn't make sense.