Add test for the EnableModules() method
Since it only allows to enable modules and not actually also disable them (the
'enable' parameter has to be 'True' according to the documentation), we can only
run the test if the modules are not already loaded. In order to have a chance to
do so, we need to try to not enable the modules before.

Also add a common method for tests to make sure the modules are enabled (if they
need that).