Add a --debug option and mark --no-debug as deprecated
With GLib logging, DEBUG and INFO messages are discarded unless the
'G_MESSAGES_DEBUG' environment variable is set to a value of "all" or
"some-domain" (for all or a specific logging domain respectively).

So if nothing extra is done on our side, DEBUG and INFO messages are discarded
which is what '--no-debug' should do. On the other hand, if we want to get DEBUG
and INFO messages, we need to add the environment variable and allow such
messages for our domain. Instead of doing this by default I believe it's better
to only turn on debug logging with an explicit '--debug' option. We can use this
option when running tests and so can all developers when debugging their changes
or anything.