Test what we can from the Drives.Ata interface
We cannot use the LIO (targetcli) devices to perform tests of this interface
because it requires some "real" SATA/SCSI devices (but disk images attached to
VMs as SATA disks seem to work). Anyway, we cannot modify such devices we find
in the system so it's only possible for us to test the read-only properties and
methods. If no devices supporting the S.M.A.R.T technology are found, all the
tests are skipped.