test_polkitd.py: Fix waiting for mock polkit to start
In the common case when we got a DBusException we previously skipped the
time.sleep() and thus were just racing through the retry loop without actually
giving the child process any time to start up.

Also raise a proper exception on timeout instead of os.abort(). The latter did
not flush stderr fast enough to actually see the failure message, and as this
happens in the parent pid we actually can raise an exception meaningfully.

LP: #1585382
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>