Use launch_spawned_job_sync() for setting file system's label
The asynchronous variant has a race condition which may cause the method call to
time out due to the spawned process exiting sooner than a callback for the
"::complete" signal is connected. And there's nothing resource-hungry to do in
the handle_set_label() function after the job is spawned so waiting with it
after the job is finished means no harm.