Use launch_spawned_job_sync() for swap operations
The asynchronous variant has a race condition which may cause the method call to
time out due to the spawned process exiting sooner than a callback for the
"::complete" signal is connected. And there's nothing resource-hungry to do in
the functions after the jobs are spawned so waiting with it after the jobs are
finished means no harm.