Build with libblockdev's bcache functionality enabled
Bcache functionality in the libblockdev-kbd plugin is now optional. If we were
using pkg-config, we would get proper cflags, but since we are using
libblockdev's plugins as standalone libraries, we need to do this on our
own. This makes sure the WITH_BD_BCACHE macro is defined which enables the
bcache functionality when building things with the libblockdev-kbd plugin.