Fix failing btrfs tests
Btrfs sucks at producing proper uevents. From the tests failures it looks like
uevents are either not produced at all on many changes or before signatures are
actually written to disk. If we synthesize our own artificial 'change' events on
devices that should get them, we suddenly start getting correct values.

This needs a real fix somewhere. Either in the code so that we somehow deal with the weird
behavior of btrfs (which is however much more work than we can spend on this
now) or somewhere in kernel/btrfs, because these issues seem to only happen with
rawhide kernels.

So let's just use this quick and dirty solution to make tests passing for now
and we will get back to the real issue here later.