Fallback to crypttab passphrase only if none given
If given a passphrase explicitly, we should use it instead of
trying to dig something up from /etc/crypttab. If we are given no
passphrase, we should report an error. But since there's no way
to pass None/NULL as a value for argument of type "s" over DBus,
we have to treat an empty string as no passphrase. However, an
empty string is a valid LUKS passphrase so we still need to
provide a way to pass it. This is now possible with the new
'keyfile_contents' optional argument we just need to mention the
trick in the documentation.

Also update the authorization checks and action descriptions
shown to users.

Please note that there's no change in behavior here because an
empty string passed as a value for the 'passphrase' argument
hasn't (ever?) been working before this change.