build: Declare use of C99 features and check for -std=gnu99 in autoconf
There are various places in the code base which use C99 features (such
as C++-style comments and inline variable declarations). Since there are
so many places, it seems intentional to use them — so configure.ac
should be changed to check for a C99-supporting compiler and use it.

Using the AC_PROG_CC_C99 macro adds -std=gnu99 to CFLAGS. We need gnu99
(not just c99) because udisks also uses some libc functions which are
standard on GNU systems, but not in POSIX (such as getgrouplist() and
getpwuid_r()).

Signed-off-by: Philip Withnall <withnall@endlessm.com>