udisksstate.c: Move log into udisks_state_set
In every case udisks_state_set returned error data for the sole
purpose of simply logging it.  Move this code into the function
so that we only have one copy of this common error logging logic.
As a side effect we don't appear to even need to check the return
value in all cases too, so we could probably remove the return value
as well.

Signed-off-by: Tony Asleson <tasleson@redhat.com>