run_tests.py: Install udev rules file
If you don't have 80-udisks2.rules installed the daemon will kind of work
but not completely and some subtle things will fail.

Additionally I refactored the copy code and also changed it so that if
there is no file that we are trying to replace to remove that file when
we are done, so that we don't litter the file system.  However, at the
moment it's commented out as we need these files for other unit tests
that also start the daemon.

Signed-off-by: Tony Asleson <tasleson@redhat.com>