Blacklist some changes for the UDisks.Client change signal (#1438232)
We shouldn't emit the "change" signal for Job objects and some
progress related properties (e.g. "SyncRate" for MDRaid). This
signal is used by gvfs to check for device changes and emitting
this signal for "progress" changes (e.g. when running RAID check)
causes it to run checks for new devices or mountpoints every
100 ms during the check.