udiskssimplejob.c: Allow NULL for message
The API states that passing NULL is acceptable, but when done we end up
down a path which causes an assertion which will trigger an abort if
G_DEBUG=fatal-criticals is enabled.  Pass an empty message when message is
NULL to avoid this.  Numerous other code paths are already calling this
function with message == "".

Signed-off-by: Tony Asleson <tasleson@redhat.com>