Also catch errors when trying to get object properties in tests
In some rare situations the DBus object we are trying to get
properties of may not yet be fully setup and it may even lack
the 'Properties' interface. The tests are set to wait for the
right value of a property so just make them able to deal with
issues when getting the value too.