Mark sleeps in iSCSI tests as subjects for fixing
The Login/Logout methods should make sure that the particular
device is added/removed to/from the system. The sleep() calls for
waiting for a session object to be created/destroyed should be
replaced by waits in the Login/Logout methods so that they are
really synchronous.

And when setting the initiator name, the value really must be
set immediately.