Fix how we create UDF file systems
As described in #338, 'mkudffs' is supposed be run in a different
way to actually create a modern UDF file system with all the
desired parameters.

Among the other things, older versions of 'mkudffs' are not able
to determine the device's block size so we need to do it on our
own which requires some extra effort.

Resolves: #338