Start even if a libblockdev plugin fails to load
Unless somebody tries to do something that actually requires the
plugin, it's not really needed. And if they do, they will get an
error like any other.

The next step is to be able to react dynamically on the
availability of the required tools/libraries/etc. But we need a
better support for that in libblockdev in the first place.