Add and use a service for cleaning up mount point directories
When udisks2 is used for mounting devices, it may need to create
the mount point directory. If unmount happens later, udisks2
cleans after itself and removes the mount point
directory. However, if the unmount happens during reboot, the
udisks2 daemon may already be terminated not having a chance to
do the cleanup. We need a different mechanism to do the job in
such cases.

systemd provides us with such a mechanism so let's make use of
it.

Resolves: rhbz#1384796
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>