Add a decorator for unstable tests
We know of some tests that are randomly failing without any
traceable issue on our side. Sometimes when there's a high load
on the testing VMs uevents get lost or similar things happen.

Unfortunately, the number of such tests is quite high which
results in most of the runs of the whole test suite marked as
failed because usually one of the problematic tests hitting its
tough luck case.

If we want the test run results to be useful at all we need to
make sure these known-to-randomly-fail tests do not have a chance
to doom the whole test run to end up being marked as failed. And
since Python's unittest module has no way to mark a test as
"don't care if it fails", we need our custom decorator for this.