Make sure the table_type is consistent in handle_create_partition
By using table_get_type_() we get a pointer to the property which
may change between being tested for one of the supported values
and later being actually used to determine what to do based on
its value. That may cause various failures, including
segmentation faults.

A particular example is when working with a GPT partition table
where the udev-blkid machinery can tell us that there's a "gpt"
partition table, but later change it to "PMBR" (maybe
temporarily) while the thread handling the
handle_create_partition() function is midway.