Wait for device to become initialized before probing it
libudev has the udev_device_get_is_initialized() method that does
this::

    Check if udev has already handled the device and has set up
    device node permissions and context, or has renamed a network
    device.

    This is only implemented for devices with a device node or
    network interfaces. All other devices return 1 here.

We should wait for such initialization to finish before we start
probing the device because otherwise we have a race condition and
may get outdated data or maybe even some temporary garbage. And
we should do this in the thread that can wait without blocking
the main loop.

Related: #448