Add an assertion method for checking a objects interfaces
Quite often in the tests we expect some object to have a specific
interface a method from which we want to call. However, sometimes
it takes quite long for the object to actually get the interface
(esp. in case of an external event causing the interface to be
added). Thus we need a method with the usual multi-try approach
that gives the object multiple chances to actually start
providing the interface.