Only check mount point prefix in the iso9660 tests
The mount point cleanup is done asynchronously so there's no
guarantee that the mount point from the previous test is going to
be cleaned up when some test is running. And if not, a new one
with a number suffix will be created. Which is totally fine and
expected, but we need to take that into account in the checks.