integration-test: don't error out on an unknown distro
The cpe name which is used to guess the distro is optional and not available on e.g Ubuntu. The returned value is currently only used to skip some problematic test on fedora 26, returning a default name/version on errors should lead to the same behaviour without erroring out on other distributions.