Set Job Progress property for FS check, repair, resize
If the libblockdev functions for FS check, repair or
resize support progress reporting, the Progress
property of a UDisksJob is set up and updated on
each report of the progress function.
The current job which is to be updated is managed
per thread and the progress function is also only
set up in libblockdev for the current thread.
The local progress report spans only the invokation
of the single libblockdev operation that the DBus
method handler calls in its dedicated thread.