vdo: Rename {Start,Activate}Volume -> {Start,Activate}VolumeByName
These two methods require VDO name as an identifier and can't take
object path to the backing device as an argument. This is a result
of the current vdo tool limitation and this fact should be reflected
and pronounced in the method names.

There's slight chance for improvement in the future vdo releases and
udisks API will reflect that once available.