Update the Partitions property on partition "add" event
We are currently updating the "Partitions" property of the
"PartitionTable" interface only as a result of a change event on
the disk. This works when actually creating the partition but not
when "adding" a disk that already has partitions (e.g. plugging in
a USB flash drive). This adds a special update call when handling
the "add" event for a new partition.