Fix backing device not set when unlocking multiple encrypted volume
TrueCrypt and VeraCrypt support multiple encryption of volumes (see
https://www.veracrypt.fr/en/Cascades.html). When such a volume is
unlocked, a separate device-mapper device is created for each encryption
layer. For triple-encryption, the dm tree looks like this:

tcrypt-1793 <dm-3> (254:3)
 └─tcrypt-1793_1 <dm-2> (254:2)
    └─tcrypt-1793_2 <dm-1> (254:1)
       └─ <loop1> (7:1)

The first two devices (dm-1 and dm-2) are not listed by udisks (because
of udev flags I think), but the actual cleartext device (dm-3) is.

When a device is unlocked, the crypto backing device property is set by
iterating through all devices and checking whether the backing device is
a child (slave). This didn't work, because the parent device (dm-1) was
not listed by udisks - and the actual cleartext device (dm-3) is not
updated because the backing device is not a direct child of it.

This patch fixes this by iterating recursively through childs if the
direct child doesn't have a udisks object, until either the backing
device is found or the number of childs is != 1 (because there is only
one child on each level in the multiple encryption dm tree).