udisksthreadedjob: Don't return early on cancellation request
The nice feature of GTask is to finish early when cancellation is requested,
leaving the job thread running, discarding result once finished. Even though
until recently this was set the 'completed' signal used to be spawned from
the job thread and the GTask callback was unused, resulting in no real effect.

Now with proper GTask usage of its completed callback this becomes a problem.
It's actually pretty dangerous as the thread job continues running uncontrolled,
potentially accessing memory not valid anymore.

It doesn't make much sense to leave operations doing some I/O running
in background in either case.

Also clarify this in the docs.