Fix reading passphrase from key file specified in crypttab
key file can contain arbitrary binary data including null byte
so we can't treat is as a "normal" string and use things
like strlen.

Note: udisksctl does this the right way and it is possible to
unlock the device using the '--key-file' option, only key files
specified in crypttab are broken.