Split out get_user_info from get_caller_uid_sync.
Previously, the function was lying. It didn't return the actual group of
the caller, but the default group associated with the user configured in
/etc/passwd (or equivalents).

I first wanted to change the function to return the gid of the caller.
The problem there is that glib doesn't provide any way to do that - one
can get a GCredentials [1] for the caller through a long chain of
function calls, but this doesn't actually contain a group id, except
through the platform-dependent return value of get_native.

I don't believe there's any problem with using the gid configured in
passwd instead of the caller for the file system permissions like the
current code does, but since this was a generic utility function that
might (or might not) be used in the future for all kinds of other
things, I believe splitting things up like this is a good idea to
prevent future misuse.

There were also quite a few callers asking for a gid without ever using
it, so this is a bit of an optimization, too. (Fewer calls to
getpwuid_r, which can be network calls.)

1: https://developer.gnome.org/gio/stable/GCredentials.html