Install a tmpfiles.d snippet to create /run/media on boot with systemd
Currently, when udisks is configured to use /run/media instead of
/media, on most operating systems, the /run/media directory is created
by udisks itself when the first mount is handled. This causes problems
when creating rootless or unprivileged OCI containers that are meant
to have access to udisks mount points because, depending on whether
something has been mounted or not after the current boot, /run/media
might be absent.

A rootless or unprivileged OCI container is created, by definition,
without requiring root privileges. Therefore, it's not possible to
create /run/media during container creation because creating a
directory under /run can only be done by the root user, and a missing
directory can't be bind mounted into the container. Moreover OCI
containers tend to be immutable, so one cannot bind in the directory
later once the container has been created. A practical example of such
a use case are Toolbox [2] containers that are meant to offer an
unprivileged development environment that's seamlessly integrated with
the host.

This problem can be solved by creating /run/media during boot with
systemd-tmpfiles by installing a snippet in tmpfiles.d [3].

Note that tmpfilesdir is defined via AC_SUBST in configure.ac and is
therefore not defined in data/tmpfiles.d/Makefile.am. Otherwise
Automake would complain about recursive references and error out.

The RPM spec file was updated as per the Fedora guidelines:
https://fedoraproject.org/wiki/Packaging:Tmpfiles.d

[1] https://www.opencontainers.org/
[2] https://github.com/debarshiray/toolbox
[3] https://www.freedesktop.org/software/systemd/man/tmpfiles.d.html

https://github.com/storaged-project/udisks/pull/641