udiskslinuxfilesystem: Explicitly remove the mountpoint after unmount
This guarantees that the mountpoint that has been previously created
by udisks is removed once the org.freedesktop.UDisks2.Filesystem.Unmount()
call returns.

Although this may be considered an incompatible change, it is
an unintrusive one, possibly only prolonging the method call by
a short time required for the cleanup iteration.

Fixes a race condition in the integration tests.