iscsi: Use proper parser to get and set initiator name
This ports the code for getting and setting initiator name to a proper
GKeyFile parser in order to handle comments, examples and multiple keys
in the config file properly. Some distributions do ship custom complex
config files by default.

Note that GKeyFile parser requires keys to always belong to some group,
for this we create a fake group in memory and strip it again when writing
the file back. This way we preserve all comments.

This commit also includes minor code cleanup and refactoring.