udisksdaemon: Fix leaking object references from the wait func waiting for object disappearance
The three udisks_daemon_wait_for_XXX() methods are taking the common
UDisksDaemonWaitFuncXXX family of callbacks. The wait funcs are often
shared among udisks_daemon_wait_for_object_sync() and
udisks_daemon_wait_for_object_to_disappear_sync() with the notion
of returning reference to an object if it exists, in conformance to the
UDisksDaemonWaitFunc documentation. However in case of
udisks_daemon_wait_for_object_sync() and
udisks_daemon_wait_for_objects_sync() the first time a valid object
is returned from the wait func, it is also returned to the original
caller who takes care of releasing the reference. In contrast to that
udisks_daemon_wait_for_object_to_disappear_sync() actually keeps the
wait func returning valid references, waiting for NULL (or timeout).

This change will release the reference each time wait func returns
a valid object when waiting for object disappearance. This allows
the use of common wait funcs for both appearance and disappearance.
It was never defined clearly and this was source of confusion and
possible memory and reference leaks.