Filesystem: ignore fstab entry if another device is mounted there
On Endless OS, all systems have the following /etc/fstab:

    LABEL=ostree / ext4  errors=remount-ro 0 0

If you insert a removable device containing a filesystem with the label
'ostree', then asks UDisks to mount it, it will attempt to mount it over
/. If the filesystem is not of type ext4, this will fail; if it is, this
will succeed, with potentially disastrous results.

There is already a check in is_in_fstab() which constructs the path to
the /dev/disk/by-label/$LABEL symlink and compares it to the device
being mounted. However, in this scenario that comparison succeeds: the
symlink is updated when the new device is detected.

To prevent this, use UDisksMountMonitor to look up the mount point
specified in fstab (using a new get_mount_for_path accessor); if an
existing mount is found, and it is a different device to the one being
(un)mounted, ignore the fstab entry.