udisksclient: Let the caller specify a D-Bus connection
Currently UDisksClient objects always talk to udisksd on the system bus.
This is fine for normal use, but unit tests may want to spin up a
private dbus-daemon using GTestDBus and have udisks use that, to isolate
the tests from the host. So this commit adds a new constructor so that a
GDBusConnection object can be passed in and used during initialization
of the UDisksClient. Also, reword a comment which seemed to imply that
UDisksClient objects are singletons (which they are not).

Before adding this feature we (in an Endless project) were setting
DBUS_SYSTEM_BUS_ADDRESS to the address of the GTestDBus daemon, to trick
udisks into using that daemon. However this led to intermittent unit
test failures when some I/O happened on the connection after the daemon
had been killed by g_test_dbus_down(), causing SIGTERM to be raised by
g_dbus_connection_real_closed(). With this patch and use of the new
constructor, the tests seem to always pass.

Fixes https://github.com/storaged-project/udisks/issues/679