Allow building without the daemon
The daemon part of udisks is specific to Linux, but the client library
and tools built on it are likely to be portable to other operating
systems because they consist of mostly wrappers of the D-Bus API. While
the official udisks daemon does not run on non-Linux systems, it is
possible for other people to create an unofficial service which is
mostly compatible with udisks for non-Linux systems.

This commit is intended to support gnome-control-center on FreeBSD.
gnome-control-center uses udisks client library to obtain information
for showing disk size in its 'About' panel. On FreeBSD, there is an
unofficial and incomplete implementation of udisks service called
'bsdisks', which is already good enough for this use case.