udisksmountmonitor: Remove caching, always read and parse the data
Maintaining cache of mounts and swaps turned out to be problematic on a high-traffic
system where the need of fresh information didn't always meet the actual cache state.
With a growing usage of UDisksMountMonitor becoming an authoritative source
of information, any async update as a result of external event often led to obsolete
information provided due to long callback chain / async processing of signals.

This commit essentially drops the traditional way of cache use. The cache is still
maintained only to detect list differences in order to send notifications out
e.g. for the cleanup thread. However any access from the API side always reads
and parses actual data. To avoid constant re-parsing of unchanged data, the input
data are checksum'ed and compared against last state.

Mount added/removed signals are emitted on a thread default main context set
at the time of object construction. Signals are emitted either as a result of
a change detected during any API call or by fd notification on the watched files.

All in all these changes should be transparent on the outside for the rest of
the udisks daemon.