Add sample mount_options.conf
An example file for configurable mount options. As it's picked by the tests
when running in-tree, all keys are commented out to not interfere with
regular operation. Builtin options are appended and commented out as
an example during the build process.

This file is purely optional and is supposed to be placed in /etc/udisks2.
Distributors are advised to ship it along with the docs somewhere
in /usr/share/doc, no need to place it in /etc unconditionally. By default,
however, this file will be installed as /etc/udisks2/mount_options.conf.example