modules: Add pkg-config files
Now that we include generated module D-Bus API GDBus symbols in libudisks2,
it's a good idea to provide respective pkg-config files so that developers
can check whether the feature has been compiled in. No extra libraries
for modules on a client side are needed at this time, the .pc files simply
point back to libudisks2.