modules: Introduce UDisksModule
This changes the modular API interface in a major way. New is the stateful
UDisksModule object that is a singleton to a module and acts as a handle.
It stores any module state that has been previously accessible through
daemon state pointers. UDisksModule constructor is failable and properly
reports module initialization failures. Points of extension are now proper
methods of UDisksModule.

Module objects and interfaces now hold reference to a UDisksModule, so does
the udisks_module_manager_get_modules() that is used across core daemon
objects. This contributes to thread safety.

Modules are made resident once initialized due to GType registration
constraints.