lsm: Add tests based on libstoragemgmt 'sim' plugin
By mocking a libstoragemgmt 'sim' plugin database we can actually cover
majority of the udisks lsm module functionality. The database needs to be
modified externally to spoof the WWN ID from autogenerated ID to a real ID
of one of our LIO iscsi volumes. That way we align both components so that
the udisks lsm module may attach extra interfaces on the chosen drive.

Separate database is created for each test run to avoid conflicts with
existing libstoragemgmt configuration.

Also reorder the test to run right after basic module tests so that the
rest of the test run hits the activated lsm drive hooks.