tests: Add basic stacked LVM + LUKS teardown tests
There are infinite possibilities of stacking various containers
on top of each other and together with calling teardown at random
level and having certain levels closed or deactivated creates
an incredibly complex test matrix. There are known issues
(see the last test) and more work on the modular side is needed
to get this right.