modules: Crash recovery on startup via a state file
Clients are supposed to call org.freedesktop.UDisks2.Manager.EnableModule()
as a greeter call to the daemon before accessing additional functionality
from modules. In case of daemon crashing it gets autostarted again on next
access, however modules are by default not loaded. Under normal circumstances
this is a transparent process to the clients and the purpose of this crash
recovery is to load previously active modules back explicitly.

A new state file is introduced with a list of active modules for the current
daemon session. Stored on a volatile location (typically /run/udisks2)
it gets automatically removed between system reboots. It's also removed upon
proper daemon exit. Should the daemon crash, the state file is detected and
picked up on next startup, loading recorded modules back.